Code that scrambles a message using the Vigenere square algorithm
This code will take a keyword and either a scrambled message or a
plaintext message (built to handle non-alpha characters and spaces) and
will convert it to the opposite. Build during my first year of
university, and some error handling was added in 2015. Just a fun side
project to test out my programming skills.

I'm also very proud of finding out that modulus of the message position
with key length gives me the key position that I want without having to
use 2 for loops.